+++ examples/main/shared/editor/EditorPlugins.tsx
@@ -357,7 +357,7 @@ export const uiSettings: UISettings = {
     dropDown: getLinkPanelDropDownConfig(),
     //placeholder: "Enter a URL here",
   },
-  disableDownload: true,
+  // disableDownload: true,
 };
 
 export const videoHandlers = {