+++ CHANGELOG.md
@@ -18,9 +18,15 @@
   </summary>
 
 ### :rocket: New Feature
+
+	
+
+- `plugin-video`
+  - [#2331](https://github.com/wix-incubator/rich-content/pull/2331) added new settings panel for custom videos with download toggle
 - `plugin-image`
   - [#2176](https://github.com/wix-incubator/rich-content/pull/2176) update image settings modal to have download and expand options using toggle buttons
-	
+
+
 ### :house: Internal
   
 - `unsupported-blocks/accordion/table/spoiler`
@@ -54,8 +60,6 @@
 
 - `plugin-spoiler`
   - [#2318](https://github.com/wix/ricos/pull/2318) filter spoilered media from preview
-
-## 8.27.6 (Mar 25, 2021)
 
 ### :nail_care: Polish